openmp: add dependency('openmp', method: 'cmake')
