+++ mesonbuild/dependencies/misc.py
@@ -61,51 +61,6 @@ class NetCDFDependency(ExternalDependency):
                 self.pcdep.append(pkgdep)
 
 
-class OpenMPDependency(ExternalDependency):
-    # Map date of specification release (which is the macro value) to a version.
-    VERSIONS = {
-        '201811': '5.0',
-        '201611': '5.0-revision1',  # This is supported by ICC 19.x
-        '201511': '4.5',
-        '201307': '4.0',
-        '201107': '3.1',
-        '200805': '3.0',
-        '200505': '2.5',
-        '200203': '2.0',
-        '199810': '1.0',
-    }
-
-    def __init__(self, environment, kwargs):
-        language = kwargs.get('language')
-        super().__init__('openmp', environment, language, kwargs)
-        self.is_found = False
-        if self.clib_compiler.get_id() == 'pgi':
-            # through at least PGI 19.4, there is no macro defined for OpenMP, but OpenMP 3.1 is supported.
-            self.version = '3.1'
-            self.is_found = True
-            self.compile_args = self.link_args = self.clib_compiler.openmp_flags()
-            return
-        try:
-            openmp_date = self.clib_compiler.get_define(
-                '_OPENMP', '', self.env, self.clib_compiler.openmp_flags(), [self], disable_cache=True)[0]
-        except mesonlib.EnvironmentException as e:
-            mlog.debug('OpenMP support not available in the compiler')
-            mlog.debug(e)
-            openmp_date = None
-
-        if openmp_date:
-            self.version = self.VERSIONS[openmp_date]
-            # Flang has omp_lib.h
-            header_names = ('omp.h', 'omp_lib.h')
-            for name in header_names:
-                if self.clib_compiler.has_header(name, '', self.env, dependencies=[self], disable_cache=True)[0]:
-                    self.is_found = True
-                    self.compile_args = self.link_args = self.clib_compiler.openmp_flags()
-                    break
-            if not self.is_found:
-                mlog.log(mlog.yellow('WARNING:'), 'OpenMP found but omp.h missing.')
-
-
 class ThreadDependency(ExternalDependency):
     def __init__(self, environment, kwargs):
         super().__init__('threads', environment, None, kwargs)
